Agrego jwt y sessionid como cookie ademas de logueo y registro de usuario

+++ backend/src/main/resources/db/migration/V1__create_campaign_table.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS campaign (
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   title VARCHAR(255) NOT NULL,
   description TEXT,
   short_description TEXT,
-  id_owner INT,
-  id_type INT,
-  id_category INT
+  id_owner BIGINT,
+  id_type BIGINT,
+  id_category BIGINT
 );
